provide api interface to servor
