+++ package.json
@@ -5,7 +5,7 @@
   "repository": "lukejacksonn/servor",
   "main": "./servor.js",
   "bin": {
-    "servor": "./servor.js"
+    "servor": "./cli.js"
   },
   "keywords": [
     "development",
@@ -15,8 +15,8 @@
     "reload"
   ],
   "scripts": {
-    "test": "node servor test index.html 8080 5000",
-    "start": "node servor"
+    "test": "node cli test index.html 8080 5000",
+    "start": "node cli"
   },
   "author": "Luke Jackson <lukejacksonn@gmail.com>",
   "license": "MIT"
